Refactored first_query_attempt into src/chains/sql_chain.py + renamed bio info chain.
