+++ src/agent/agent_main.py
@@ -5,10 +5,10 @@ from langchain_core.tools import Tool
 from langchain_openai import ChatOpenAI
 from pydantic import BaseModel, Field
 from langchain.tools import tool
-from chains.first_query_attempt import get_chain
+from chains.stats_sql_chain import get_chain
 from shot_maps.shot_map_plotting import goal_map_scatter_get, shot_map_scatter_get
 from chains.rag_chain import get_cba_information, get_rules_information
-from chains.bio_info_query import get_bio_chain
+from chains.bio_info_chain import get_bio_chain
 
 
 class goal_map_scatter_schema(BaseModel):
